Loop (#54)
* updating header typings and function signatures

* updating memory mw modules

* updating platform TUs

* updating runtime TUs

* updating runtime TUrelocating pragmas

* game camera positions setup cleanups

* cam in runtime type fixes

* addressing heavy regression symbols

+++ src/game/reactions.c
@@ -321,7 +321,7 @@ void plyr_spawn_anim(void* animation);
 extern int f_fatality_was_done;
 extern int g_drone_blocking_in_reaction;
 extern int g_drone_faked_out;
-extern unsigned long mode_of_play;
+extern int mode_of_play;
 static float r_complete_ermac_slam(void);
 static float r_face3_onback(void);
 void set_ani_speed(float speed);
